# docs/api/api-spec.json
{
    "openapi": "3.0.3",
    "info": {
        "title": "sartre API",
        "description": "A Hacker News-style API built with Hono, featuring user authentication, posts, comments, and voting functionality.",
        "version": "1.0.0",
        "contact": {
            "name": "sartre",
            "url": "http://localhost:3001"
        }
    },
    "servers": [
        {
            "url": "http://localhost:3000/api",
            "description": "Development server"
        }
    ],
    "paths": {
        "/user": {
            "get": {
                "summary": "Get current user",
                "description": "Returns the currently authenticated user's information",
                "tags": ["Authentication"],
                "security": [
                    {
                        "cookieAuth": []
                    }
                ],
                "responses": {
                    "200": {
                        "description": "User information retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "type": "object",
                                            "properties": {
                                                "username": {
                                                    "type": "string",
                                                    "example": "johndoe"
                                                }
                                            }
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Not authenticated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/posts": {
            "get": {
                "summary": "Get posts",
                "description": "Retrieve a paginated list of posts with optional filtering and sorting",
                "tags": ["Posts"],
                "parameters": [
                    {
                        "name": "limit",
                        "in": "query",
                        "description": "Number of posts per page",
                        "schema": {
                            "type": "integer",
                            "default": 10,
                            "minimum": 1,
                            "maximum": 100
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Page number",
                        "schema": {
                            "type": "integer",
                            "default": 1,
                            "minimum": 1
                        }
                    },
                    {
                        "name": "sortBy",
                        "in": "query",
                        "description": "Sort posts by",
                        "schema": {
                            "type": "string",
                            "enum": ["points", "recent"],
                            "default": "points"
                        }
                    },
                    {
                        "name": "order",
                        "in": "query",
                        "description": "Sort order",
                        "schema": {
                            "type": "string",
                            "enum": ["asc", "desc"],
                            "default": "desc"
                        }
                    },
                    {
                        "name": "author",
                        "in": "query",
                        "description": "Filter by author username",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "site",
                        "in": "query",
                        "description": "Filter by site domain",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Posts retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "allOf": [
                                        {
                                            "$ref": "#/components/schemas/SuccessResponse"
                                        },
                                        {
                                            "type": "object",
                                            "properties": {
                                                "data": {
                                                    "type": "array",
                                                    "items": {
                                                        "$ref": "#/components/schemas/Post"
                                                    }
                                                },
                                                "pagination": {
                                                    "$ref": "#/components/schemas/Pagination"
                                                }
                                            }
                                        }
                                    ]
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "summary": "Create new post",
                "description": "Create a new post (requires authentication)",
                "tags": ["Posts"],
                "security": [
                    {
                        "cookieAuth": []
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "example": "Interesting Article Title"
                                    },
                                    "url": {
                                        "type": "string",
                                        "format": "uri",
                                        "example": "https://example.com/article"
                                    },
                                    "content": {
                                        "type": "string",
                                        "example": "This is the content of the post..."
                                    }
                                },
                                "required": ["title"]
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Post created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "allOf": [
                                        {
                                            "$ref": "#/components/schemas/SuccessResponse"
                                        },
                                        {
                                            "type": "object",
                                            "properties": {
                                                "data": {
                                                    "type": "object",
                                                    "properties": {
                                                        "postId": {
                                                            "type": "integer",
                                                            "example": 123
                                                        }
                                                    }
                                                }
                                            }
                                        }
                                    ]
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Authentication required",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid input",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "cookieAuth": {
                "type": "apiKey",
                "in": "cookie",
                "name": "better-auth.session_token",
                "description": "Session cookie for authentication"
            }
        },
        "schemas": {
            "SuccessResponse": {
                "type": "object",
                "properties": {
                    "success": {
                        "type": "boolean",
                        "example": true
                    },
                    "message": {
                        "type": "string",
                        "example": "Operation completed successfully"
                    }
                },
                "required": ["success", "message"]
            },
            "ErrorResponse": {
                "type": "object",
                "properties": {
                    "success": {
                        "type": "boolean",
                        "example": false
                    },
                    "error": {
                        "type": "string",
                        "example": "An error occurred"
                    },
                    "isFormError": {
                        "type": "boolean",
                        "example": false
                    }
                },
                "required": ["success", "error"]
            },
            "Pagination": {
                "type": "object",
                "properties": {
                    "page": {
                        "type": "integer",
                        "example": 1
                    },
                    "totalPages": {
                        "type": "integer",
                        "example": 10
                    }
                },
                "required": ["page", "totalPages"]
            },
            "Post": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 123
                    },
                    "title": {
                        "type": "string",
                        "example": "Interesting Article Title"
                    },
                    "url": {
                        "type": "string",
                        "format": "uri",
                        "example": "https://example.com/article",
                        "nullable": true
                    },
                    "content": {
                        "type": "string",
                        "example": "This is the content of the post...",
                        "nullable": true
                    },
                    "points": {
                        "type": "integer",
                        "example": 42
                    },
                    "createdAt": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-01-15T10:30:00Z"
                    },
                    "commentCount": {
                        "type": "integer",
                        "example": 15
                    },
                    "author": {
                        "type": "object",
                        "properties": {
                            "id": {
                                "type": "string",
                                "example": "user123"
                            },
                            "username": {
                                "type": "string",
                                "example": "johndoe"
                            }
                        },
                        "required": ["id", "username"]
                    },
                    "isUpvoted": {
                        "type": "boolean",
                        "example": false
                    }
                },
                "required": [
                    "id",
                    "title",
                    "points",
                    "createdAt",
                    "commentCount",
                    "author",
                    "isUpvoted"
                ]
            },
...
        }
    }
}